+++ src/App.js
@@ -3,6 +3,8 @@ import {
   DocumentTreeComponent,
   EditorComponent,
 } from "./Component/index.js";
+import { documentLinkClickEvent } from "./events/index.js";
+import { route } from "./router/route.js";
 import { DocumentTree, Document } from "./domain/index.js";
 import { request } from "./api.js";
 import { getItem, setItem, removeItem } from "./storage/storage.js";
@@ -26,11 +28,12 @@ export default class App extends Component {
             action: "click",
             tag: "a",
             target: "a",
-            callback: async ({ event }) => {
-              event.preventDefault();
-
-              this.route({ url: event.target.href });
-            },
+            callback: route,
+            // callback: async ({ event, target }) => {
+            // documentLinkClickEvent(eventProps);
+            //   event.preventDefault();
+            //   route({ url: event.target.href });
+            // },
           },
           {
             action: "click",
@@ -193,42 +196,42 @@ export default class App extends Component {
     });
   }
 
-  async route({ url }) {
-    if (this.editor.state.id !== -1) {
-      const { title, content } = this.editor.state;
-      this.saveDocumentToServer({ title, content });
-    }
-    history.pushState(null, null, url);
-    const urlSplit = url.split("/");
-    const [routeName, documentId] = urlSplit.slice(
-      urlSplit.length - 2,
-      urlSplit.length
-    );
-    switch (routeName) {
-      case "documents":
-        const savedDocument = new Document(await this.getDocument(documentId));
-        try {
-          const { title, content, tmpSaveDate } = getItem(
-            "documents/" + documentId
-          );
-          if (tmpSaveDate > savedDocument.updatedAt) {
-            if (confirm("임시저장된 데이터가 있습니다. 불러오시겠습니까?")) {
-              this.editor.state = savedDocument.cloneNewDocument({
-                title,
-                content,
-                updatedAt: tmpSaveDate,
-              });
-              return;
-            }
-          }
-        } catch (error) {
-          console.log(
-            "임시저장된 데이터가 없습니다. 서버에서 데이터를 불러옵니다."
-          );
-        }
-        this.editor.state = savedDocument;
-        removeItem("documents/" + documentId);
-        break;
-    }
-  }
+  // async route({ url }) {
+  //   if (this.editor.state.id !== -1) {
+  //     const { title, content } = this.editor.state;
+  //     this.saveDocumentToServer({ title, content });
+  //   }
+  //   history.pushState(null, null, url);
+  //   const urlSplit = url.split("/");
+  //   const [routeName, documentId] = urlSplit.slice(
+  //     urlSplit.length - 2,
+  //     urlSplit.length
+  //   );
+  //   switch (routeName) {
+  //     case "documents":
+  //       const savedDocument = new Document(await this.getDocument(documentId));
+  //       try {
+  //         const { title, content, tmpSaveDate } = getItem(
+  //           "documents/" + documentId
+  //         );
+  //         if (tmpSaveDate > savedDocument.updatedAt) {
+  //           if (confirm("임시저장된 데이터가 있습니다. 불러오시겠습니까?")) {
+  //             this.editor.state = savedDocument.cloneNewDocument({
+  //               title,
+  //               content,
+  //               updatedAt: tmpSaveDate,
+  //             });
+  //             return;
+  //           }
+  //         }
+  //       } catch (error) {
+  //         console.log(
+  //           "임시저장된 데이터가 없습니다. 서버에서 데이터를 불러옵니다."
+  //         );
+  //       }
+  //       this.editor.state = savedDocument;
+  //       removeItem("documents/" + documentId);
+  //       break;
+  //   }
+  // }
 }
