+++ src/events/documentTreeEvent.js
@@ -1,0 +1,4 @@
+export const documentLinkClickEvent = async ({ event }) => {
+  event.preventDefault();
+  this.route({ url: event.target.href });
+};
