+++ src/events/index.js
@@ -1,0 +1,1 @@
+export { documentLinkClickEvent } from "./documentTreeEvent.js";
